OrthoDraw/Bends: conforming IntType for uniform_int_distribution
[rand.req.genl]/1.5 restricts a uniform_int_distribution's IntType to
short/int/long/long long (+ unsigned); Int8 (std::int8_t) is not among them,
so uniform_int_distribution<Int8> is undefined under C++20. Widen the draw to
int, matching the analogous small-range draw in OrthoDraw/TurnRegularize.hpp:423.

No behavior change: the draw is {-1,0,1} and is immediately copy-assigned to an
Int8 on the next line. make all builds clean; Tier 1 stays 2495/2495.

Prepared with Claude Code (Opus 4.8)

+++ src/OrthoDraw/Bends.hpp
@@ -161,7 +161,7 @@ void RandomizeBends_impl(
 {
     using CrossingMatrix_T = PD_T::C_Arcs_T;
     
-    std::uniform_int_distribution<Int8> dice( -1, 1 );
+    std::uniform_int_distribution<int> dice( -1, 1 );
     
     const Int c_count = pd.MaxCrossingCount();
